exit + kill(in progress)

+++ loup-garou/game.js
@@ -28,7 +28,7 @@ function center_button(){
 function player_pressed(player){
     switch (step){
         case 0:
-            console.log(JSON.parse(getCookie("players_list"))[player-1]);
+            console.log(players_list[player-1]);
             break;
         case 1:
             switch (night){
@@ -46,13 +46,16 @@ function player_pressed(player){
     }
 }
 
+/**
+ * @return {string} Return who won (false if nobody)
+ */
 function is_game_finished(){
     let village_winning = true;
     let wolves_winning = true;
     let alive = [];
     players_list.forEach(player => {
         if (player.role == "wolf" || player.is_infected) village_winning = false;
-        else wolves_winning = false;
+        else if (player.role) wolves_winning = false;
         if (player.role){
             alive.push(player.name);
         }
@@ -63,24 +66,48 @@ function is_game_finished(){
     return false;
 }
 
+/**
+ * @param {number} nb_player - player to kill
+ * @param {boolean} try_married - kill the player married to them (true)
+ * @return {string} The new role of the player
+ */
 function kill(nb_player, try_married = true){
     let player = players_list[nb_player-1]
     if (player.lifes > 0){
-        for (let i = 0; i > 20; i++){
-            player.role = "new-role";
-            if (! is_game_finished){
-                break;
-            }
+        player.lifes --;
+        for (let i = 0; i <= 30; i++){
+            player.role = "---new_role---";
+            if (! is_game_finished) break;
         }
-    }else player.role = false;
-    game_end(is_game_finished());
-    if (player.married_to && try_married){
-        kill(player.married_to,false);
+    }else{
+        player.role = false;
     }
+    players_list[nb_player-1] = player;
+    game_end(is_game_finished());
+    if (player.married_to){
+        if (try_married) kill(player.married_to,false);
+        else save_players();
+        player.married_to = 0
+    }else save_players();
+    return player.role;
 }
 
 function game_end(winner){
     if (winner){
         console.log(winner+" ont gagnés !");
     }else return ;
+}
+
+/**
+ * @param {object} list - input array
+ * @return {} A random object in the array
+ */
+function random(list){
+    if (typeof list === "object") return list[Math.floor(Math.random()*list.length)];
+    return ;
+}
+
+function save_players(){
+    setCookie("players_list",JSON.stringify(players_list));
+    return JSON.parse(getCookie("players_list"));
 }
